Fix flake8 errors.
